# swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/user":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}}},"summary":"Create a new user","tags":["User"]},"get":{"operationId":"findAll","parameters":[{"name":"skip","required":true,"in":"query","schema":{"minimum":0,"default":0,"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindAllUserDto"}}}}},"summary":"Get all users","tags":["User"]}},"/api/user/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Get a user by ID","tags":["User"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Update a user by ID","tags":["User"]},"delete":{"operationId":"delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Delete a user by ID","tags":["User"]}},"/api/patient":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePatientDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePatientDto"}}}}},"summary":"Create a new patient","tags":["Patient"]}},"/api/patient/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginationDto"}}}}},"summary":"Get a patient by ID","tags":["Patient"]}},"/api/media/upload":{"post":{"operationId":"uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadFileDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaDto"}}}}},"tags":["Media"]}},"/api/media/{id}":{"get":{"operationId":"get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaDto"}}}}},"tags":["Media"]}},"/api/Plate":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"summary":"Create a new plate","tags":["Plate"]},"get":{"operationId":"findAll","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"fistName","required":false,"in":"query","schema":{"type":"string"}},{"name":"lastName","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"userId","required":false,"in":"query","schema":{"type":"number"}},{"name":"skip","required":true,"in":"query","schema":{"minimum":0,"default":0,"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindAllPlateDto"}}}}},"summary":"Get Plates by filters","tags":["Plate"]}},"/api/Plate/Process":{"post":{"operationId":"process","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultDto"}}}}},"summary":"Start Processing","tags":["Plate"]}},"/api/Plate/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlateDto"}}}}},"summary":"Get a Plate by ID","tags":["Plate"]}},"/api/Plate/result/{id}":{"get":{"operationId":"findOneResult","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlateDto"}}}}},"summary":"Get a Plate by ID","tags":["Plate"]}},"/api/Plate/result/{id}/reject":{"patch":{"operationId":"rejectResult","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlateDto"}}}}},"tags":["Plate"]}},"/api/Plate/result/{id}/accept":{"patch":{"operationId":"acceptResult","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlateDto"}}}}},"tags":["Plate"]}},"/api/Plate/result/{id}/confuse":{"patch":{"operationId":"confuseResult","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlateDto"}}}}},"tags":["Plate"]}},"/api/antibiotic":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAntiBioticDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AntiBioticDto"}}}}},"summary":"Create a new plate","tags":["Antibiotic"]}},"/api/antibiotic/{id}":{"get":{"operationId":"get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AntiBioticDto"}}}}},"summary":"Create a new plate","tags":["Antibiotic"]}},"/api/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/refresh":{"post":{"operationId":"refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}}},"info":{"title":"Competition","description":"An API to manage competitions","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","in":"header","name":"Authorization","description":"Enter your Bearer token"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"}},"required":["email","phoneNumber","name","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"email":{"type":"string"}},"required":["id","name","createdAt","updatedAt","email"]},"FindAllUserDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}},"required":["data"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"},"name":{"type":"string"}}},"CreatePatientDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"number":{"type":"string"}},"required":["firstName","lastName","number"]},"PaginationDto":{"type":"object","properties":{"skip":{"type":"number","minimum":0,"default":0},"limit":{"type":"number","minimum":1,"default":10}},"required":["skip","limit"]},"UploadFileDto":{"type":"object","properties":{"media":{"type":"string","format":"binary"}},"required":["media"]},"MediaDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"path":{"type":"string"}},"required":["id","name","path"]},"CreatePlateDto":{"type":"object","properties":{"patientFirstName":{"type":"string"},"patientLastName":{"type":"string"},"patientNumber":{"type":"string"},"mediaId":{"type":"number"},"notes":{"type":"string"}},"required":["patientFirstName","patientLastName","patientNumber","mediaId","notes"]},"ResultDto":{"type":"object","properties":{"id":{"type":"number"},"plateId":{"type":"number"},"status":{"type":"string"},"antibioticDetections":{"type":"array","items":{"type":"string"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","plateId","status","antibioticDetections","createdAt","updatedAt"]},"PatientDto":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"number":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","firstName","lastName","number","createdAt","updatedAt"]},"PlateDto":{"type":"object","properties":{"id":{"type":"number"},"patient":{"$ref":"#/components/schemas/PatientDto"},"image":{"$ref":"#/components/schemas/MediaDto"},"notes":{"type":"string"},"result":{"$ref":"#/components/schemas/ResultDto"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","patient","image","notes","result","createdAt","updatedAt"]},"FindAllPlateDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/PlateDto"}}},"required":["data"]},"CreateAntiBioticDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"AntiBioticDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["id","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"}},"required":["email","name","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]}}},"security":[{"bearer":[]}]}